fix: Finance widget is not loading
